packages/watchmen-agent-cli/pyproject.toml
[tool.poetry]
name = "agent-cli"
version = "16.0.0"
description = "Watchmen agent friendly CLI for topic and pipeline synchronization"
authors = ["imma-team"]
packages = [
    { include = "agent_cli", from = "src" }
]

[tool.poetry.dependencies]
python = "~3.12"
requests = "^2.32.3"
pyyaml = "^6.0.3"

[tool.poetry.group.dev.dependencies]

[tool.poetry.scripts]
agent-cli = "agent_cli.main:run"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
